+++ release-notes.rst
@@ -4,6 +4,19 @@ chevah-brink release notes
 Here are the release notes for past brink version.
 
 
+Version 0.69.4, 2018-06-23
+--------------------------
+
+* Use fixed chmod for sync on Windows.
+* Add SHA256 checksum file.
+
+
+Version 0.69.3, 2018-05-23
+--------------------------
+
+* Fix rsync command line.
+
+
 Version 0.69.2, 2018-05-23
 --------------------------